pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srebench"
version = "1.0.0"
description = "SREBench — On-Call Incident Response OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "pydantic>=2.7.1",
    "PyYAML>=6.0.1",
    "numpy>=1.26.4",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
baseline = ["openai>=1.35.0"]
dev = ["pytest>=8.2.2", "pytest-asyncio>=0.23.7"]

[tool.setuptools.packages.find]
include = ["app*", "client*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
